+++ port_range/__init__.py
@@ -28,7 +28,7 @@ PY3 = sys.version_info == 3
 if PY3:
     basestring = (str, bytes)
 
-__version__ = '1.1.0'
+__version__ = '2.0.0'
 
 
 class PortRange(object):